+++ backend/localrep/models/__init__.py
@@ -1,4 +1,6 @@
 from .algo import Algo
+from .algo import AlgoInput
+from .algo import AlgoOutput
 from .computeplan import ComputePlan
 from .computetask import ComputeTask
 from .datamanager import DataManager
@@ -9,6 +11,8 @@ from .performance import Performance
 
 __all__ = [
     "Algo",
+    "AlgoInput",
+    "AlgoOutput",
     "ComputePlan",
     "ComputeTask",
     "DataManager",