+++ backend/localrep/models/algo.py
@@ -2,8 +2,41 @@ from django.contrib.postgres.fields import ArrayField
 from django.db import models
 
 import orchestrator.algo_pb2 as algo_pb2
+import orchestrator.common_pb2 as common_pb2
 from localrep.models.utils import AssetPermissionMixin
 from localrep.models.utils import URLValidatorWithOptionalTLD
+
+
+class AlgoInput(models.Model):
+    class Kind(models.TextChoices):
+        ASSET_DATA_SAMPLE = common_pb2.AssetKind.Name(common_pb2.ASSET_DATA_SAMPLE)
+        ASSET_DATA_MANAGER = common_pb2.AssetKind.Name(common_pb2.ASSET_DATA_MANAGER)
+        ASSET_MODEL = common_pb2.AssetKind.Name(common_pb2.ASSET_MODEL)
+
+    algo = models.ForeignKey("Algo", on_delete=models.deletion.CASCADE, related_name="inputs")
+    identifier = models.CharField(max_length=100)
+    kind = models.CharField(max_length=64, choices=Kind.choices)
+    optional = models.BooleanField(default=False)
+    multiple = models.BooleanField(default=False)
+
+    class Meta:
+        unique_together = (("algo", "identifier"),)
+        ordering = ["identifier"]  # default order for relations serializations
+
+
+class AlgoOutput(models.Model):
+    class Kind(models.TextChoices):
+        ASSET_MODEL = common_pb2.AssetKind.Name(common_pb2.ASSET_MODEL)
+        ASSET_PERFORMANCE = common_pb2.AssetKind.Name(common_pb2.ASSET_PERFORMANCE)
+
+    algo = models.ForeignKey("Algo", on_delete=models.deletion.CASCADE, related_name="outputs")
+    identifier = models.CharField(max_length=100)
+    kind = models.CharField(max_length=64, choices=Kind.choices)
+    multiple = models.BooleanField(default=False)
+
+    class Meta:
+        unique_together = (("algo", "identifier"),)
+        ordering = ["identifier"]  # default order for relations serializations
 
 
 class Algo(models.Model, AssetPermissionMixin):