+++ backend/substrapp/tests/factory.py
@@ -51,6 +51,8 @@ import uuid
 from django.utils import timezone
 
 from localrep.models import Algo
+from localrep.models import AlgoInput
+from localrep.models import AlgoOutput
 from localrep.models import ComputePlan
 from localrep.models import ComputeTask
 from localrep.models import DataManager
@@ -62,6 +64,47 @@ from localrep.models.computetask import TaskDataSamples
 DEFAULT_OWNER = "MyOrg1MSP"
 DEFAULT_CHANNEL = "mychannel"
 DUMMY_CHECKSUM = "dummy-checksum"
+
+
+# This logic belongs to the SDK but we replicate the mapping to generate realistic assets
+ALGO_INPUTS_PER_CATEGORY = {
+    "ALGO_SIMPLE": {
+        "datasamples": {"kind": "ASSET_DATA_SAMPLE", "multiple": True, "optional": False},
+        "model": {"kind": "ASSET_MODEL", "multiple": False, "optional": True},
+        "opener": {"kind": "ASSET_DATA_MANAGER", "multiple": False, "optional": False},
+    },
+    "ALGO_AGGREGATE": {
+        "model": {"kind": "ASSET_MODEL", "multiple": True, "optional": False},
+    },
+    "ALGO_COMPOSITE": {
+        "datasamples": {"kind": "ASSET_DATA_SAMPLE", "multiple": True, "optional": False},
+        "local": {"kind": "ASSET_MODEL", "multiple": False, "optional": True},
+        "opener": {"kind": "ASSET_DATA_MANAGER", "multiple": False, "optional": False},
+        "shared": {"kind": "ASSET_MODEL", "multiple": False, "optional": True},
+    },
+    "ALGO_METRIC": {
+        "datasamples": {"kind": "ASSET_DATA_SAMPLE", "multiple": True, "optional": False},
+        "opener": {"kind": "ASSET_DATA_MANAGER", "multiple": False, "optional": False},
+        "predictions": {"kind": "ASSET_MODEL", "multiple": False, "optional": False},
+    },
+}
+
+
+ALGO_OUTPUTS_PER_CATEGORY = {
+    "ALGO_SIMPLE": {
+        "model": {"kind": "ASSET_MODEL", "multiple": False},
+    },
+    "ALGO_AGGREGATE": {
+        "model": {"kind": "ASSET_MODEL", "multiple": False},
+    },
+    "ALGO_COMPOSITE": {
+        "local": {"kind": "ASSET_MODEL", "multiple": False},
+        "shared": {"kind": "ASSET_MODEL", "multiple": False},
+    },
+    "ALGO_METRIC": {
+        "performance": {"kind": "ASSET_PERFORMANCE", "multiple": False},
+    },
+}
 
 
 def get_storage_address(asset: str, key: str, field: str) -> str:
@@ -154,7 +197,8 @@ def create_algo(
 ) -> Algo:
     if key is None:
         key = uuid.uuid4()
-    return Algo.objects.create(
+
+    algo = Algo.objects.create(
         key=key,
         name=name,
         category=category,
@@ -168,6 +212,13 @@ def create_algo(
         channel=channel,
         **get_permissions(owner, public),
     )
+
+    for identifier, algo_input in ALGO_INPUTS_PER_CATEGORY[category].items():
+        AlgoInput.objects.create(algo=algo, identifier=identifier, **algo_input)
+    for identifier, algo_output in ALGO_OUTPUTS_PER_CATEGORY[category].items():
+        AlgoOutput.objects.create(algo=algo, identifier=identifier, **algo_output)
+
+    return algo
 
 
 def create_datamanager(